+++ src/logging/WriterBackend.h
@@ -70,7 +70,7 @@ protected:
      * indicate what happened, and the writer and its thread will eventually
      * be terminated.
      */
-    virtual bool WriteLogs(int num_writes, threading::Value*** vals) override final;
+    virtual bool WriteLogs(size_t num_writes, threading::Value*** vals) override final;
     
     virtual bool RunHeartbeat(double network_time, double current_time) override final;
